+++ settings.h
@@ -1,0 +1,18 @@
+/*
+ * beanbricks.c: a questionable breakout clone in C and Raylib.
+ *
+ * Copyright (c) Eason Qin <eason@ezntek.com>, 2024-2025.
+ *
+ * This source code form is wholly licensed under the MIT/Expat license. View
+ * the full license text in the root of the project.
+ *
+ * INFO: settings screen
+ */
+
+#ifndef _SETTINGS_H
+#define _SETTINGS_H
+
+void draw_settings(void);
+void update_settings(void);
+
+#endif
