+++ titlescreen.h
@@ -1,0 +1,22 @@
+/*
+ * beanbricks.c: a questionable breakout clone in C and Raylib.
+ *
+ * Copyright (c) Eason Qin <eason@ezntek.com>, 2024-2025.
+ *
+ * This source code form is wholly licensed under the MIT/Expat license. View
+ * the full license text in the root of the project.
+ *
+ * INFO: titlescreen logic
+ */
+
+#ifndef _TITLESCREEN_H
+#define _TITLESCREEN_H
+
+void draw_titlescreen(void);
+void draw_titlescreen_gui(void);
+
+void update_titlescreen(void);
+
+void reset_titlescreen(void);
+
+#endif
